requirements.txt
django~=3.2.9
django-environ>=0.8.1
django-cleanup>=5.2.0
django-filter>=21.1.0
django-import-export>=2.6.1
django-request-logging>=0.7.3
django-storages[azure]>=1.12.3
whitenoise>=5.3.0
djangorestframework>=3.12.4
djangorestframework-simplejwt>=4.8.0
djoser>=2.1.0
psycopg2-binary>=2.9.2
gunicorn>=20.1.0
